feat : add userId in to Jwt, create jwtUserDetails to fetch userId and other data, updated api's to use auth token to fetch userId

+++ backend/src/main/java/com/ccms/portal/service/CardApplicationService.java
@@ -2,9 +2,15 @@ package com.ccms.portal.service;
 
 
 import com.ccms.portal.dto.request.CardApplicationRequest;
+import com.ccms.portal.dto.request.CreateCardRequest;
 import com.ccms.portal.enums.CardApplicationStatus;
 import com.ccms.portal.model.CardApplicationEntity;
 import com.ccms.portal.repository.CardApplicationRepository;
+import com.ccms.portal.util.JwtUserDetails;
+import com.ccms.portal.util.JwtUtil;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,21 +18,39 @@ import java.util.Optional;
 
 @Service
 public class CardApplicationService {
+    private final JwtUtil jwtUtil;
     private final CardApplicationRepository repository;
+    private final CardService cardService;
 
-    public CardApplicationService(CardApplicationRepository repository){
+    public CardApplicationService(CardApplicationRepository repository, JwtUtil jwtUtil, CardService cardService){
+        this.jwtUtil = jwtUtil;
         this.repository = repository;
+        this.cardService = cardService;
     }
 
     public CardApplicationEntity apply(CardApplicationRequest application){
+        JwtUserDetails currentUser = (JwtUserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+
+        Long userId = currentUser.getUserId();
         CardApplicationEntity applicationEntity = new CardApplicationEntity();
         applicationEntity.setStatus(CardApplicationStatus.Pending);
         applicationEntity.setCardTypeId(application.getCardTypeId());
-        applicationEntity.setUserId(application.getUserId());
+        applicationEntity.setRequestedLimit(application.getRequestLimit());
+        applicationEntity.setUserId(userId);
         return repository.save(applicationEntity);
     }
 
-    public List<CardApplicationEntity> getApplications(Long userId){
+    public List<CardApplicationEntity> getApplications(){
+
+        JwtUserDetails currentUser = (JwtUserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+
+        Long userId = currentUser.getUserId();
         return repository.findAllByUserId(userId);
     }
     public Optional<CardApplicationEntity> getApplicationById(Long id){
@@ -34,12 +58,35 @@ public class CardApplicationService {
     }
 
     public CardApplicationEntity updateStatus(Long id, CardApplicationStatus request){
+        JwtUserDetails currentUser = (JwtUserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+        Long userId = currentUser.getUserId();
         CardApplicationEntity application = repository.findById(id).orElseThrow(()-> new RuntimeException("Application not found"));
+        if(!application.getUserId().equals(userId)){
+            throw new RuntimeException("You are not authorized to update this application");
+        }
+        CreateCardRequest createCardRequest = new CreateCardRequest();
+        createCardRequest.setCardTypeId(application.getCardTypeId());
+        createCardRequest.setUserId(application.getUserId());
+        createCardRequest.setCreditLimit(application.getRequestedLimit()*1.0);
+        cardService.createCard(createCardRequest);
         application.setStatus(request);
         return repository.save(application);
     }
 
     public void deleteApplication(Long id) {
+        JwtUserDetails currentUser = (JwtUserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+        Long userId = currentUser.getUserId();
+        CardApplicationEntity application  = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Card application with id " + id + " not found"));
+        if (!application.getUserId().equals(userId)) {
+            throw new RuntimeException("You are not authorized to delete this application");
+        }
         repository.deleteById(id);
     }
 }
